refactor: move oidctoken into sdks/stack so the provider can use it
The Terraform provider consumes sdks/stack and cannot import packages
under bins/cli/internal. Moving it into the SDK module also keeps the
dependency edge pointing root -> sdk rather than dragging the root
module into a provider build.
