+++ bins/cli/cmd/cli.go
@@ -15,7 +15,6 @@ import (
 
 	"github.com/nuonco/nuon/bins/cli/internal/agentmode"
 	"github.com/nuonco/nuon/bins/cli/internal/config"
-	"github.com/nuonco/nuon/bins/cli/internal/oidctoken"
 	"github.com/nuonco/nuon/bins/cli/internal/services/actions"
 	"github.com/nuonco/nuon/bins/cli/internal/services/apps"
 	"github.com/nuonco/nuon/bins/cli/internal/services/auth"
@@ -33,6 +32,7 @@ import (
 	"github.com/nuonco/nuon/bins/cli/internal/services/variables"
 	"github.com/nuonco/nuon/bins/cli/internal/services/version"
 	"github.com/nuonco/nuon/bins/cli/internal/ui"
+	"github.com/nuonco/nuon/sdks/stack/oidctoken"
 )
 
 type cli struct {
